requirements.txt
# Web Framework
Flask==3.1.0
Werkzeug==3.1.3

# Environment
python-dotenv==1.1.0

# Database
pymongo==4.12.1
dnspython==2.7.0

# AI / OpenAI
openai==1.77.0
pydantic==2.11.4

# HTTP Clients
requests==2.32.3
httpx==0.28.1

# PDF Generation
reportlab==4.4.0

# AFIP Integration
pysimplesoap==1.16.2

# Template Engine (included with Flask)
Jinja2==3.1.6
MarkupSafe==3.0.2
